normalize page views
This normalizes the pageviews and makes them independent from the
configured rewriting and useslash settings. It also creates virtual page
views for search and admin. The search views are strucutred in a way
that makes it easy to track the internal search key words.

The normalization currently strips out all query parameters. It has to
be decided if that is desired.